+++ grpc/interceptor/interceptor.go
@@ -4,7 +4,6 @@ import (
 	"io"
 	"os"
 
-	middlewarecommon "github.com/Azure/aks-middleware/common"
 	"github.com/Azure/aks-middleware/grpc/client/mdforward"
 	"github.com/Azure/aks-middleware/grpc/common"
 	"github.com/Azure/aks-middleware/grpc/common/autologger"
@@ -147,6 +146,6 @@ func DefaultServerInterceptors(options ServerInterceptorLogOptions) []grpc.Unary
 			logging.WithFieldsFromContext(common.GetFields),
 		),
 		responseheader.UnaryServerInterceptor(httpcommon.MetadataToHeader),
-		recovery.UnaryServerInterceptor(middlewarecommon.GetRecoveryOpts()...),
+		recovery.UnaryServerInterceptor(common.GetRecoveryOpts()...),
 	}
 }
